[nextest-runner] move some internal events into another module
Add `runner/internal_events.rs`, and move some types into this module.
